web/requirements.txt
django
djangorestframework
django-oauth-toolkit
django-rest-auth
django-extensions
django-filter

oauthlib
requests_oauthlib

fabric
aiohttp
psycopg2
redis
mockredis
